--- c_bindings/crunch64.h
@@ -1,9 +1,0 @@
-#ifndef CRUNCH64_H
-#define CRUNCH64_H
-#pragma once
-
-#include "crunch64_error.h"
-#include "yay0.h"
-#include "yaz0.h"
-
-#endif
