+++ c_bindings/include/crunch64.h
@@ -1,0 +1,9 @@
+#ifndef CRUNCH64_H
+#define CRUNCH64_H
+#pragma once
+
+#include "crunch64/error.h"
+#include "crunch64/yay0.h"
+#include "crunch64/yaz0.h"
+
+#endif
